requirements.txt
streamlit
gspread
google-auth
pandas
plotly
xhtml2pdf
